Move plugin action classes to a sub-package
This makes their use as library classes more obvious.
